+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Full documentation for rocFFT is available at [rocfft.readthedocs.io](https://ro
 - Improved performance of 1D lengths < 2048 that use Bluestein's algorithm.
 - Reduced time for generating code during plan creation.
 - Optimized 3D R2C/C2R lengths 32, 84, 128.
+- Optimized batched small 1D R2C/C2R cases.
 
 ### Added
 - Added gfx1101 to default AMDGPU_TARGETS.